feat: support custom cluster root path
