+++ Earthfile
@@ -55,7 +55,7 @@ VERSION:
 
     COPY . ./
 
-    RUN echo $(git describe --exact-match --tags || echo "v0.0.0-$(git log --oneline -n 1 | cut -d" " -f1)") > VERSION
+    RUN echo $(git describe --exact-match --tags || echo "v0.0.0-$(git rev-parse --short=8 HEAD)") > VERSION
 
     SAVE ARTIFACT VERSION VERSION
