+++ pkg/provider/mount.tmpl
@@ -1,0 +1,12 @@
+[Unit]
+Description={{.Name}} mount unit
+Before=local-fs.target k3s.service k3s-agent.service
+
+[Mount]
+What={{.Source}}
+Where={{.Target}}
+Type=none
+Options=bind
+
+[Install]
+WantedBy=local-fs.target
