+++ pkg/types/mount.go
@@ -1,0 +1,7 @@
+package types
+
+type MountPoint struct {
+	Name   string
+	Source string
+	Target string
+}
