+++ pkg/types/rke2.go
@@ -1,0 +1,7 @@
+package types
+
+type RKE2Config struct {
+	Token  string   `yaml:"token"`
+	Server string   `yaml:"server"`
+	TLSSan []string `yaml:"tls-san"`
+}
